requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Supabase Integration
supabase==2.3.4

# CORS & Security
django-cors-headers==4.3.1
django-filter==23.5

# Authentication
djangorestframework-simplejwt==5.3.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Email & SMS
sendgrid==6.11.0
twilio==8.10.3

# Payments
requests==2.31.0

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Image Processing
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Development
ipython==8.18.1

# Admin Interface
django-admin-interface==0.32.0

